fix: booster packs not showing what items were gained after purchase
