+++ src/services/inventoryService.ts
@@ -28,12 +28,10 @@ import {
 } from "../types/requestTypes";
 import { logger } from "@/src/utils/logger";
 import { getWeaponType, getExalted } from "@/src/services/itemDataService";
-import { getRandomWeightedReward } from "@/src/services/rngService";
 import { ISyndicateSacrifice, ISyndicateSacrificeResponse } from "../types/syndicateTypes";
 import { IEquipmentClient, IItemConfig } from "../types/inventoryTypes/commonInventoryTypes";
 import {
     ExportArcanes,
-    ExportBoosterPacks,
     ExportCustoms,
     ExportFlavour,
     ExportGear,
@@ -180,21 +178,6 @@ export const addItem = async (
                 FlavourItems: [await addCustomization(typeName, accountId)]
             }
         };
-    }
-    if (typeName in ExportBoosterPacks) {
-        const pack = ExportBoosterPacks[typeName];
-        const InventoryChanges = {};
-        for (const weights of pack.rarityWeightsPerRoll) {
-            const result = getRandomWeightedReward(pack.components, weights);
-            if (result) {
-                logger.debug(`booster pack rolled`, result);
-                combineInventoryChanges(
-                    InventoryChanges,
-                    (await addItem(accountId, result.type, result.itemCount)).InventoryChanges
-                );
-            }
-        }
-        return { InventoryChanges };
     }
     if (typeName in ExportUpgrades || typeName in ExportArcanes) {
         const inventory = await getInventory(accountId);
